// src/app/models/ichild-profile.ts
export interface IChildProfile {
    id: string;
    childName: string;
    fatherName: string;
    fatherPhoneNumber: string;
    fatherEmail: string;
    motherName: string;
    motherPhoneNumber: string;
    motherEmail: string;
    address: string;
    note: string;
    updatedTime: string;
    createdTime: string;
  }